chore: API Gateway에 팝업 서비스 라우팅 및 Swagger config 파일 변경
