+++ popi-api-gateway/src/main/resources/application-local.yml
@@ -62,6 +62,23 @@ spring:
             - RewritePath=/reservations/(?<segment>.*), /$\{segment}
             - JwtAuthenticationFilter
 
+        - id: popup-docs
+          uri: lb://POPUPS
+          predicates:
+            - Path=/popups/v3/api-docs
+          filters:
+            - RemoveRequestHeader=Cookie
+            - RewritePath=/popups/(?<segment>.*), /$\{segment}
+        - id: popup-service
+          uri: lb://POPUPS
+          predicates:
+            - Path=/popups/**
+            - Method=GET,POST,DELETE
+          filters:
+            - RemoveRequestHeader=Cookie
+            - RewritePath=/popups/(?<segment>.*), /$\{segment}
+            - JwtAuthenticationFilter
+
       globalcors:
         cors-configurations:
           "[/**]":
@@ -93,6 +110,9 @@ springdoc:
     urls[2]:
       name: 예약 서비스
       url: /reservations/v3/api-docs
+    urls[3]:
+      name: 팝업 서비스
+      url: /popups/v3/api-docs
 
 jwt:
   access-token-secret: ${JWT_ACCESS_TOKEN_SECRET}
