--- popi-popup-service/src/main/java/com/lgcns/popipopupservice/config/SwaggerConfig.java
@@ -1,45 +1,0 @@
-package com.lgcns;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import io.swagger.v3.oas.models.servers.Server;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(
-                        new Info()
-                                .title("PoPI Popup Server API")
-                                .description("PoPI 팝업 서비스 API 명세서입니다.")
-                                .version("v0.0.1"))
-                .addServersItem(new Server().url("/popups"))
-                .components(authSetting())
-                .addSecurityItem(securityRequirement());
-    }
-
-    private Components authSetting() {
-        return new Components()
-                .addSecuritySchemes(
-                        "accessToken",
-                        new SecurityScheme()
-                                .type(SecurityScheme.Type.HTTP)
-                                .scheme("bearer")
-                                .bearerFormat("JWT")
-                                .in(SecurityScheme.In.HEADER)
-                                .name("Authorization"));
-    }
-
-    private SecurityRequirement securityRequirement() {
-        return new SecurityRequirement().addList("accessToken");
-    }
-}
